исправил директория presenter, перенес все render() из ьmain.js в board-presenter.js
