+++ src/main.js
@@ -1,13 +1,7 @@
-import FilterView from './view/filter-view.js';
-import SortView from './view/sort-view.js';
-import {render} from './render.js';
-import BoardPresenter from './view/presenter/board-presenter.js';
+import BoardPresenter from './presenter/board-presenter.js';
 
 const filtersElement = document.querySelector('.trip-controls__filters');
 const tripElement = document.querySelector('.trip-events');
-const boardPresenter = new BoardPresenter({boardContainer: tripElement});
-
-render(new FilterView(), filtersElement);
-render(new SortView(), tripElement);
+const boardPresenter = new BoardPresenter({boardContainer: tripElement, filterContainer: filtersElement});
 
 boardPresenter.init();
